Everything except nave and artifacts

+++ src/consoleUI/Menu.java
@@ -1,10 +1,16 @@
 package consoleUI;
 
+import engine.Exceptions.XmlPathContainsNonRepositoryObjectsException;
+import engine.Exceptions.XmlRepositoryAlreadyExistsException;
 import engine.Manager;
 
+import javax.print.DocFlavor;
+import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
+import java.util.StringTokenizer;
 
 public class Menu {
     private ArrayList<MenuItem> menuItems;
@@ -12,7 +18,7 @@ public class Menu {
     private Manager magitManager;
 
     public Menu() {
-        this.menuItems = new ArrayList<MenuItem>();
+        this.menuItems = new ArrayList<>();
         this.scanner = new Scanner(System.in);
         magitManager = new Manager();
         buildMenu();
@@ -47,7 +53,7 @@ public class Menu {
             printMenuItems();
             option = scanner.nextLine();
 
-            if (option.equals("13")) {
+            if (option.equals("14")) {
                 quit = true;
             }
             try {
@@ -72,7 +78,36 @@ public class Menu {
         addMenuItem("2", "Load repository from XML", new Runnable() {
             @Override
             public void run() {
-                //magitManager.Load repository();
+                String fileFullPath;
+                String choice;
+                System.out.println("Enter the XML file's full path:");
+                fileFullPath = scanner.nextLine();
+                try {
+                    magitManager.ValidateXMLRepository(fileFullPath);
+                    System.out.println("The xml file was loaded successfully");
+                } catch (XmlRepositoryAlreadyExistsException ex) {
+                    System.out.println("The path is already a repository. please choose one of the following options: \r\n" +
+                            "[1] replace the existing repository with the one loaded from the xml file\r\n" +
+                            "[2] continue working with the existing repository");
+                    choice = getUserChoiceBetweenOneAndTwo();
+                    if (choice.equals("1")) {
+                        try {
+                            magitManager.createRepositoryFromMagitRepository();
+                            System.out.println("The xml file was loaded successfully");
+                        } catch (Exception ex2) {
+                            System.out.println("Error occurred while creating the repository:");
+                            System.out.println(ex2.getMessage());
+                        }
+                    } else { // the user chose to stay with the existing repository
+                        System.out.println("The xml file was not loaded");
+                    }
+
+                } catch (XmlPathContainsNonRepositoryObjectsException e) {
+                    System.out.println(e.getMessage());
+                } catch (Exception e) {
+                    System.out.println("Error occurred while loading the file:");
+                    System.out.println(e.getMessage());
+                }
             }
         });
 
@@ -96,6 +131,11 @@ public class Menu {
         addMenuItem("4", "Show all files of current commit", new Runnable() {
             @Override
             public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
                 List<String> allFilesData = null;
                 if (!magitManager.commitsWereExecuted()) {
                     System.out.println("No commits were executed in this repository");
@@ -117,20 +157,38 @@ public class Menu {
         addMenuItem("5", "Show status", new Runnable() {
             @Override
             public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
                 System.out.println(magitManager.GetRepositoryDetails());
-                System.out.println(magitManager.GetStatus());
+                try {
+                    System.out.println(magitManager.GetStatus());
+                } catch (IOException e) {
+                    System.out.println("Error: runTime failure");
+                }
             }
         });
 
         addMenuItem("6", "Commit", new Runnable() {
             @Override
             public void run() {
-                if (!magitManager.thereAreUncommittedChanges()) {
-                    System.out.println("There are no open changes");
-                } else {
-                    System.out.println("Enter a message for the commit:");
-                    String message = scanner.nextLine();
-                    magitManager.ExecuteCommit(message);
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
+                try {
+                    if (!magitManager.thereAreUncommittedChanges()) {
+                        System.out.println("There are no open changes");
+                    } else {
+                        System.out.println("Enter a message for the commit:");
+                        String message = scanner.nextLine();
+                        magitManager.ExecuteCommit(message);
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
                 }
             }
         });
@@ -138,10 +196,19 @@ public class Menu {
         addMenuItem("7", "Show all branches", new Runnable() {
             @Override
             public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
                 List<String> allBranchesDetails;
-                allBranchesDetails = magitManager.GetAllBranchesDetails();
-                for (String str : allBranchesDetails) {
-                    System.out.println(str);
+                try {
+                    allBranchesDetails = magitManager.GetAllBranchesDetails();
+                    for (String str : allBranchesDetails) {
+                        System.out.println(str);
+                    }
+                } catch (IOException e) {
+                    System.out.println("Error: runTime failure");
                 }
             }
         });
@@ -149,16 +216,18 @@ public class Menu {
         addMenuItem("8", "create new branch", new Runnable() {
             @Override
             public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
                 String branchName;
                 String choice = "";
                 System.out.println("Enter the new branch name:");
                 branchName = scanner.nextLine();
                 System.out.println("Would you like to checkout " + branchName + "?\r\n[1] yes   [2] no");
-                choice = scanner.nextLine();
-                while (!choice.equals("1") && !choice.equals("2")) {
-                    System.out.println("Invalid choice. Please enter '1' or '2'");
-                    choice = scanner.nextLine();
-                }
+                choice = getUserChoiceBetweenOneAndTwo();
+
                 try {
                     magitManager.CreateNewBranch(branchName, choice.equals("1"));
                 } catch (Exception e) {
@@ -170,11 +239,17 @@ public class Menu {
         addMenuItem("9", "Delete branch", new Runnable() {
             @Override
             public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
                 String branchName;
                 System.out.println("Enter the name of the branch you would like to delete:");
                 branchName = scanner.nextLine();
                 try {
                     magitManager.DeleteBranch(branchName);
+                    System.out.println("The branch " + branchName + " was deleted");
                 } catch (Exception e) {
                     System.out.println(e.getMessage());
                 }
@@ -184,6 +259,11 @@ public class Menu {
         addMenuItem("10", "Checkout", new Runnable() {
             @Override
             public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
                 String branchToCheckout;
                 System.out.println("Enter the name of the branch you would like to checkout: ");
                 branchToCheckout = scanner.nextLine();
@@ -192,11 +272,8 @@ public class Menu {
                     if (magitManager.thereAreUncommittedChanges()) {
                         System.out.println("There are uncommitted changes");
                         System.out.println("Would you like to commit before checkout " + branchToCheckout + "?\r\n[1] yes   [2] no");
-                        choice = scanner.nextLine();
-                        while (!choice.equals("1") && !choice.equals("2")) {
-                            System.out.println("Invalid choice. Please enter '1' or '2'");
-                            choice = scanner.nextLine();
-                        }
+                        choice = getUserChoiceBetweenOneAndTwo();
+
                         if (choice.equals("1")) {
                             runCommand("6");
                         }
@@ -208,16 +285,23 @@ public class Menu {
             }
         });
 
-
         addMenuItem("11", "Show history of active branch", new Runnable() {
             @Override
             public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
                 String activeBranchHistory;
-                activeBranchHistory = magitManager.GetActiveBranchHistory();
-                System.out.println(activeBranchHistory);
+                try {
+                    activeBranchHistory = magitManager.GetActiveBranchHistory();
+                    System.out.println(activeBranchHistory);
+                } catch (IOException e) {
+                    System.out.println("Error: runTime failure");
+                }
             }
         });
-
 
         addMenuItem("12", "Create new repository", new Runnable() {
             @Override
@@ -236,12 +320,60 @@ public class Menu {
             }
         });
 
+        addMenuItem("13", "Reset head branch", new Runnable() {
+            @Override
+            public void run() {
+                if (!magitManager.HasActiveRepository()) {
+                    System.out.println("Invalid option, the system currently is not working on any repository\r\n" +
+                            "Please choose repository to work on, by creating,loading or switching to one");
+                    return;
+                }
+                String sha1;
+                String choice;
+                try {
+                    if (magitManager.thereAreUncommittedChanges()) {
+                        System.out.println("There are uncommitted changes");
+                        System.out.println("Would you like to commit before reset?\r\n[1] yes   [2] no");
+                        choice = getUserChoiceBetweenOneAndTwo();
 
-        addMenuItem("13", "Quit", new Runnable() {
+                        if (choice.equals("1")) {
+                            try {
+                                runCommand("6");
+                            } catch (Exception e) {
+                                System.out.println(e.getMessage());
+                                return;
+                            }
+                        }
+                    }
+                } catch (IOException e) {
+                    System.out.println("Error: runTime failure");
+                    return;
+                }
+                System.out.println("Enter the sha1 of the commit you would like to reset the head branch to: ");
+                sha1 = scanner.nextLine();
+                try {
+                    magitManager.ResetHeadBranch(sha1);
+                } catch (Exception e) {
+                    System.out.println(e.getMessage());
+                }
+            }
+        });
+
+        addMenuItem("14", "Quit", new Runnable() {
             @Override
             public void run() {
                 System.out.println("Goodbye!");
             }
         });
+    }
+
+    private String getUserChoiceBetweenOneAndTwo() {
+        String choice;
+        choice = scanner.nextLine();
+        while (!choice.equals("1") && !choice.equals("2")) {
+            System.out.println("Invalid choice. Please enter '1' or '2'");
+            choice = scanner.nextLine();
+        }
+        return choice;
     }
 }
